Improved reusability of code by adding the tile.dart file
